fix: :adhesive_bandage: localstorage key conflict with new game
Because wordle+ is also hosted on github pages and both games have a key called mode a key conflic occured leading to a null pointer. I fixed it by prefixing all keys for this game with cg-. I also changed the resizing header to a pure css solution, hid the start button when it flips around on non timed gamemodes and improved the input filtering on the various number inputs in settings. I also made the storing and retrieving of values from localstorage less brittle.